add editing a room functionality

+++ src/services/apiCabins.js
@@ -7,7 +7,6 @@ export async function getCabins() {
     console.error(error);
     throw new Error("Cabins could not be loaded");
   }
-
   return data;
 }
 
@@ -17,36 +16,48 @@ export async function deleteCabin(id) {
     console.error(error);
     throw new Error("Cabins could not be deleted");
   }
-
   return data;
 }
 
-export async function createCabin(newCabin) {
-  const imageName = `${Math.random()}-${newCabin.image.name}`.replaceAll(
-    "/",
-    ""
-  );
-  const imagePath = `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
-  const { data, error } = await supabase
-    .from("cabins")
-    .insert([{ ...newCabin, image: imagePath }])
-    .select();
+export async function createEditCabin(newCabin, id) {
+  // Check if the image is a string (URL) or a file object
+  const isString = typeof newCabin?.image === "string";
+  const hasImagePath = isString && newCabin.image.startsWith(supabaseUrl);
+
+  // Generate imageName if newCabin.image is a file object
+  const imageName = isString
+    ? ""
+    : `${Math.random()}-${newCabin.image.name}`.replaceAll("/", "");
+
+  const imagePath = hasImagePath
+    ? newCabin.image
+    : `${supabaseUrl}/storage/v1/object/public/cabin-images/${imageName}`;
+
+  // 1. Create/edit cabin
+  let query = supabase.from("cabins");
+  if (!id) query = query.insert([{ ...newCabin, image: imagePath }]);
+  if (id) query = query.update({ ...newCabin, image: imagePath }).eq("id", id);
+
+  const { data, error } = await query.select().single();
 
   if (error) {
     console.error(error);
-    throw new Error("Cabins could not be created");
+    throw new Error("Cabins could not be created or updated");
   }
 
-  const { error: storageError } = await supabase.storage
-    .from("cabin-images")
-    .upload(imageName, newCabin.image);
+  // Only upload the image if it's a File object
+  if (!isString) {
+    const { error: storageError } = await supabase.storage
+      .from("cabin-images")
+      .upload(imageName, newCabin.image);
 
-  if (storageError) {
-    await supabase.from("cabins").delete().eq("id", data.id);
-    console, error(storageError);
-    throw new Error(
-      "Cabins image could not be uploaded and the cabin was not created"
-    );
+    if (storageError) {
+      await supabase.from("cabins").delete().eq("id", data.id);
+      console.error(storageError);
+      throw new Error(
+        "Cabins image could not be uploaded and the cabin was not created"
+      );
+    }
   }
 
   return data;
